Bugfix CDDD and main application to address ripple effects from refactoring.
