+++ benchmark/cuchembm/__main__.py
@@ -48,12 +48,11 @@ def wait_for_megamolbart_service(inferrer):
     return False
 
 
-def save_metric_results(metric_list, output_dir):
+def save_metric_results(mode_name, metric_list, output_dir):
     metric_df = pd.concat(metric_list, axis=1).T
     log.info(metric_df)
     metric = metric_df['name'].iloc[0].replace(' ', '_')
-    iteration = metric_df['iteration'].iloc[0]
-    csv_file_path = os.path.join(output_dir, f'{metric}_iteration{iteration}.csv')
+    csv_file_path = os.path.join(output_dir, f'{mode_name}_{metric}.csv')
     write_header = False if os.path.exists(csv_file_path) else True
     metric_df.to_csv(csv_file_path, index=False, mode='a', header=write_header)
 
@@ -189,7 +188,7 @@ def main(cfg):
                 result['gene'] = kwargs['gene']
 
             result_list.append(result)
-        save_metric_results(result_list, output_dir)
+        save_metric_results(cfg.model.name, result_list, output_dir)
 
 
 if __name__ == '__main__':
