+++ benchmark/scripts/benchmark.yaml
@@ -11,7 +11,7 @@ root:
   handlers: [console]
 
 model:
-  name: MegaMolBART
+  name: CDDD
   params:
     url: localhost:50051
 
@@ -19,7 +19,6 @@ metric:
   validity:
     enabled: True
     radius:
-      - 1
       - 2
 
   unique:
@@ -45,7 +44,7 @@ metric:
 
 sampling:
   max_seq_len: 512
-  input_size: 20000
+  input_size: 5000
   sample_size: 10
 
 output:
