+++ cuchem/cuchem/wf/cluster/__init__.py
@@ -2,10 +2,12 @@ from typing import List
 
 import cupy
 import numpy
-from cuchemcommon.data.helper.chembldata import ADDITIONAL_FEILD, IMP_PROPS
+
 from cuml.metrics import pairwise_distances
+
+from cuchemcommon.data.helper.chembldata import ADDITIONAL_FEILD, IMP_PROPS
 from cuchem.utils.distance import tanimoto_calculate
-from cuchem.utils.metrics import spearmanr
+from cuchemcommon.utils.metrics import spearmanr
 
 
 class BaseClusterWorkflow:
